requirements.txt
aiohttp==3.12.15
numpy==2.3.3
python-dotenv==1.1.1
quart==0.20.0
quart_cors==0.8.0
Requests==2.32.5
